Review.java
The totalSenitment method finds the sentiment value of each word. Letters are then added together and then stops when the method reaches a space. A method is then called that finds the sentiment value of the entire word. Next the value is stored and added into sentiment, finally the value is returned by totalSentiment.